% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_convolution.R
\name{brm_convolution}
\alias{brm_convolution}
\title{Convolution brm Wrappers}
\usage{
brm_convolution(
  formula,
  data,
  conv_mean = c(2.5, 1),
  conv_sd = c(1, 0.5),
  conv_max = 30,
  conv_varying = "fixed",
  hold_out_time = 28,
  dry_run = FALSE,
  ...
)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data.frame that must contain the date, location (as loc), primary
(the data that the outcome is a convolution of) and at least the outcome as
specifed in \code{formula}.}

\item{conv_mean}{Numeric vector defining the prior for the log mean of the convolution
(mean and standard deviation).}

\item{conv_sd}{Numeric vector defining the prior for the standard deviation of the
convolution (mean and standard deviation).}

\item{conv_max}{Integer defining the maximum index to use for the convolution.}

\item{conv_varying}{A character string defining the variability of the convolution
distribution. By default this is fixed by other supported options include allowing
the convolution distribution to vary across location (using a random effect).}

\item{hold_out_time}{Integer, number of days to hold out from the likelihood. This
is useful as initially the outcome will depend on primary data outside of the range of
the training set.}

\item{dry_run}{Logical, defaults to TRUE. For testing purposes should just the \code{stan}
code be output with not fitting done.}

\item{...}{Additional parameters passed to \code{brms::brm}.}
}
\value{
A ""brmsfit" object.
}
\description{
A wrapper for \code{brms::brm} that implements a custom model where a
secondary indicator is predicted by a primary indicator over some convolution and
scaled by a some fraction. The fraction can then be modelled using the regression
framework of \code{brms}.
}
\author{
Sam Abbott
}
