% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{brmid.brmsid_convolution}
\alias{brmid.brmsid_convolution}
\title{Delay Convolution Model}
\usage{
\method{brmid}{brmsid_convolution}(
  formula = ~1,
  conv_mean = ~1,
  conv_sd = ~1,
  family = negbinomial(),
  data,
  priors,
  id_stancode,
  use_default_formula = TRUE,
  dry = FALSE,
  ...
)
}
\arguments{
\item{formula}{A model formula}

\item{conv_mean}{Formula for the convolution mean. Defaults to intercept
only.}

\item{conv_sd}{Formula for the convolution standard deviation. Defaults to
intercept only.}

\item{data}{A data.frame as produced by \code{prepare} that must contain the date,
location (as loc), primary (the data that the outcome is a convolution of)
and at least the outcome as specifed in \code{formula}.}

\item{...}{Additional parameters passed to \code{brms::brm}.}
}
\value{
A "brmsfit" object or stan code (if \code{dry = TRUE}).
}
\description{
A model that assumes that a secondary observations can be
predicted using a convolution of a primary observation multipled by some
scaling factor. An example use case of this model is to estimate the
case fatality rate (with the primary observation being cases and the
secondary observation being deaths) and then explore factors that influence
it.
}
\author{
Sam Abbott
}
