% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsid-methods.R
\name{brmid}
\alias{brmid}
\title{Infectious disease modelling wrapper for brm}
\usage{
brmid(formula, data, ...)
}
\arguments{
\item{formula}{Formula of the main regression equation for the \code{brmid} model.}

\item{data}{A data frame as prepared for modelling using \code{prepare}.}
}
\description{
Infectious disease modelling wrapper for brm
}
\author{
Sam Abbott
}
