% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expose_idbrms_stan_fns}
\alias{expose_idbrms_stan_fns}
\title{Expose package stan functions in R}
\usage{
expose_idbrms_stan_fns(files, dir, ...)
}
\arguments{
\item{files}{A character vector indicating the target files}

\item{dir}{A character string indicating the directory for the file}

\item{...}{Additional arguments passed to \code{rstan::expose_stan_functions}.}
}
\description{
This function exposes internal stan functions in R from a user
supplied list of target files. Allows for testing of stan functions in R and
potentially user use in R code.
}
