% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{id_priors.brmsid_convolution}
\alias{id_priors.brmsid_convolution}
\title{Define priors for the delay convolution model}
\usage{
\method{id_priors}{brmsid_convolution}(data, conv_type = "lognormal")
}
\description{
Define priors for the delay convolution model
}
