% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{id_priors.idbrms_convolution}
\alias{id_priors.idbrms_convolution}
\title{Define priors for the delay convolution model}
\usage{
\method{id_priors}{idbrms_convolution}(
  data,
  scale = c(round(log(0.1), 2), 0.05),
  cmean = c(2.5, 1),
  lcsd = c(-0.5, 0.25),
  ...
)
}
\arguments{
\item{data}{A data.frame as produced by \code{prepare} that must contain the date,
location (as loc), primary (the data that the outcome is a convolution of)
and secondary (the observation of interest. Should have class
"idbrms_convolution".}

\item{scale}{Vector of length two defining the mean and the standard
deviation of the normal prior for the scaling factor.}

\item{cmean}{Vector of length two defining the mean and standard deviation of
the log mean of the delay distribution.}

\item{lcsd}{Vector of length two defining the mean and standard deviation of
the log standard deviation logged.
the standard deviation to be greater than 0 on the unconstrained scale.}

\item{...}{Additional arguments for method.}
}
\description{
Define priors for the delay convolution model
}
