% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{id_priors}
\alias{id_priors}
\title{Define model specific priors}
\usage{
id_priors(data, ...)
}
\arguments{
\item{data}{A data frame as prepared for modelling using \code{prepare} with a
class associated with the model prepared for.}

\item{...}{Additional arguments for method.}
}
\description{
Define model specific priors
}
\examples{
x <- 1
class(x) <- "idbrms_convolution"
id_priors(x)
}
\author{
Sam Abbott
}
