% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{id_stancode}
\alias{id_stancode}
\title{Define model specific stancode}
\usage{
id_stancode(data, ...)
}
\arguments{
\item{data}{A data frame as prepared for modelling using \code{prepare} with a
class associated with the model prepared for.}

\item{...}{Additional arguments for method.}
}
\description{
Define model specific stancode
}
\author{
Sam Abbott
}
