% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idbrms_stan_chunk}
\alias{idbrms_stan_chunk}
\title{Read in a idbrms Stan code chunk}
\usage{
idbrms_stan_chunk(path)
}
\arguments{
\item{path}{The path within the "stan" folder of the installed idbrms
package to the stan code chunk of interest.}
}
\value{
A character string containing the stan code chunk of interest.
}
\description{
Read in a idbrms Stan code chunk
}
\examples{
idbrms_stan_chunk("functions/idbrms_convolve.stan")
}
