% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{idbrms_version_stanvar}
\alias{idbrms_version_stanvar}
\title{Label a idbrms stan model with a version indicator}
\usage{
idbrms_version_stanvar()
}
\value{
A brms stanvar chunk containing the package version used to build
the stan code.
}
\description{
Label a idbrms stan model with a version indicator
}
