% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{prepare.brmsid_convolution}
\alias{prepare.brmsid_convolution}
\title{Prepare data for fitting using a convolution model}
\usage{
\method{prepare}{brmsid_convolution}(data, location, primary, initial_obs = 14, max_convolution = 30)
}
\arguments{
\item{location}{Character string, variable to use as the spatial location.}

\item{primary}{Character string, variable to use as the primary observation.}

\item{initial_obs}{Integer, number of observations to hold out from the
likelihood. This is useful as initially the outcome will depend on primary
data outside of the range of the training set and including this could bias
the estimated delay distribution. Defaults to 14 days.}

\item{max_convolution}{Integer defining the maximum index to use for the
convolution. Defaults to 30 days.}
}
\description{
Prepare data for fitting using a convolution model
}
\examples{
df <- data.frame(
   region = "France", deaths = 1:10, cases = 1:10,
   date = seq(as.Date("2020-10-01"), by = "days", length.out = 10)
   )

df <- prepare(
  df, model = "convolution", location = "region", primary = "cases"
  )
print(df)
}
\author{
Sam Abbott
}
