% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsid-methods.R
\name{prepare}
\alias{prepare}
\alias{prepare.default}
\title{Prepare data for modelling with brmid}
\usage{
prepare(data, ...)

\method{prepare}{default}(data, model, ...)
}
\arguments{
\item{...}{Additional arguments passed to model specific prepare functions}

\item{model}{Character string, model type to prepare to use. Support options
are "convolution".}
}
\description{
Prepare data for modelling with brmid

Default method used when preparing data
}
\examples{
df <- data.frame(
   region = "France", deaths = 1:10, cases = 1:10,
   date = seq(as.Date("2020-10-01"), by = "days", length.out = 10)
   )

df <- prepare(
  df, model = "convolution", location = "region", primary = "cases"
  )
print(df)
}
\author{
Sam Abbott

Sam Abbott
}
