% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{idbrm}
\alias{idbrm}
\title{Interface for infectious disease modelling using brms.}
\usage{
idbrm(data, formula, family, priors, custom_stancode, dry = FALSE, ...)
}
\arguments{
\item{data}{A data frame as prepared for modelling using \code{prepare} with a
class associated with the model prepared for.}

\item{formula}{A formula as defined using \code{id_formula} or as supported by
\code{brms::brm}.}

\item{family}{A observation model family as defined in \code{brms}.}

\item{priors}{A list of priors as defined using \code{brms} or \code{id_priors}.
Defaults to the the \code{id_priors} defined for the model class being fit.}

\item{custom_stancode}{A list of \code{stanvars} used to define custom stancode in
\code{brms}. By default uses the code designed for the model class being fit (as
specified using \code{id_stancode}).}

\item{dry}{Logical, defaults to TRUE. For testing purposes should just the \code{stan}
code be output with not fitting done.}

\item{...}{Additional arguments to pass to \code{brms::brm}.}
}
\description{
Interface for infectious disease modelling using brms.
}
\examples{
\donttest{
# define some example data
library(data.table)
dt <- data.table(
   region = "France", cases = seq(10, 500, by = 10),
   date = seq(as.Date("2020-10-01"), by = "days", length.out = 50)
   )
dt[, deaths := as.integer(shift(cases, 5) * 0.1)]
dt[is.na(deaths), deaths := 0]

dt <- prepare(
  dt, model = "convolution", location = "region", 
  primary = "cases", secondary = "deaths",
  )

# fit the convolution model using a Poisson observation model
fit <- idbrm(data = dt, family = poisson(link = "identity"))
}
}
\author{
Sam Abbott
}
