% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R, R/methods.R
\name{id_formula.idbrms_convolution}
\alias{id_formula.idbrms_convolution}
\alias{id_formula}
\title{Define a formula for the convolution model}
\usage{
\method{id_formula}{idbrms_convolution}(data, scale = ~1, cmean = ~1, lcsd = ~1, ...)

id_formula(data, ...)
}
\arguments{
\item{data}{A data.frame as produced by \code{prepare} that must contain the date,
location (as loc), primary (the data that the outcome is a convolution of)
and secondary (the observation of interest. Should have class
"idbrms_convolution".}

\item{scale}{Formula for the scaling of primary observations to secondary
observations.}

\item{cmean}{Formula for the convolution mean. Defaults to intercept
only.}

\item{lcsd}{Vector of length two defining the mean and standard deviation of
the log standard deviation logged.
the standard deviation to be greater than 0 on the unconstrained scale.}

\item{...}{Additional arguments to pass to \code{brms::brm}.}

\item{csd}{Formula for the convolution standard deviation. Defaults to
intercept only.}
}
\description{
Define a formula for the convolution model

Define a model specific formula
}
\author{
Sam Abbott

Sam Abbott
}
