% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{id_stancode.idbrms_convolution}
\alias{id_stancode.idbrms_convolution}
\title{Define stan code for a delay convolution model}
\usage{
\method{id_stancode}{idbrms_convolution}(data)
}
\description{
Define stan code for a delay convolution model
}
\examples{
x <- 1
class(x) <- "brmsid_convolution"
custom_stan <- id_stancode(x)
}
