% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmsid-methods.R
\name{id_stancode}
\alias{id_stancode}
\title{Define model specific stancode}
\usage{
id_stancode(data, ...)
}
\description{
Define model specific stancode
}
\examples{
x <- 1
class(x) <- "brmsid_convolution"
custom_stan <- id_stancode(x)
}
\author{
Sam Abbott
}
