% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{prepare.brmsid_convolution}
\alias{prepare.brmsid_convolution}
\title{Prepare data for fitting using a convolution model}
\usage{
\method{prepare}{brmsid_convolution}(
  data,
  location,
  primary,
  secondary,
  initial_obs = 14,
  max_convolution = 30
)
}
\arguments{
\item{data}{A data frame containing at least two integer observations and a date
variable.}

\item{location}{Character string, variable to use as the spatial location.}

\item{primary}{Character string, variable to use as the primary observation.}

\item{secondary}{Character string, variable to use as the secondary
observation.}

\item{initial_obs}{Integer, number of observations to hold out from the
likelihood. This is useful as initially the outcome will depend on primary
data outside of the range of the training set and including this could bias
the estimated delay distribution. Defaults to 14 days.}

\item{max_convolution}{Integer defining the maximum index to use for the
convolution. Defaults to 30 days.}
}
\description{
Prepare data for fitting using a convolution model
}
\examples{
# define some example data
library(data.table)
dt <- data.table(
   region = "France", cases = seq(10, 500, by = 10),
   date = seq(as.Date("2020-10-01"), by = "days", length.out = 50)
   )
dt[, deaths := as.integer(shift(cases, 5) * 0.1)]
dt[is.na(deaths), deaths := 0]

dt <- prepare(
  dt, model = "convolution", location = "region", 
  primary = "cases", secondary = "deaths",
  )
dt[]
}
\author{
Sam Abbott
}
