% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{prepare}
\alias{prepare}
\alias{prepare.default}
\title{Prepare data for modelling with idbrm}
\usage{
prepare(data, ...)

\method{prepare}{default}(data, model, ...)
}
\arguments{
\item{data}{A dataframe to be used for modelling}

\item{...}{Additional arguments passed to model specific prepare functions}

\item{model}{Character string, model type to prepare to use.
Supported options are "convolution".}
}
\description{
Prepare data for modelling with idbrm

Default method used when preparing data
}
\examples{
# define some example data
library(data.table)
dt <- data.table(
   region = "France", cases = seq(10, 500, by = 10),
   date = seq(as.Date("2020-10-01"), by = "days", length.out = 50)
   )
dt[, deaths := as.integer(shift(cases, 5) * 0.1)]
dt[is.na(deaths), deaths := 0]

dt <- prepare(
  dt, model = "convolution", location = "region",
  primary = "cases", secondary = "deaths",
  )
dt[]
}
\author{
Sam Abbott

Sam Abbott
}
